add build workspace
