+++ packages/fuels/Cargo.toml
@@ -19,7 +19,6 @@ fuels-signers = { version = "0.33.0", path = "../fuels-signers" }
 fuels-test-helpers = { version = "0.33.0", path = "../fuels-test-helpers" }
 fuels-types = { version = "0.33.0", path = "../fuels-types" }
 
-
 [dev-dependencies]
 anyhow = "1.0.58"
 ctor = " 0.1"
