+++ packages/fuels/tests/Forc.toml
@@ -1,0 +1,71 @@
+[workspace]
+members = [
+  "types/b512",
+  "types/call_empty_return",
+  "types/complex_types_contract",
+  "types/contract_output_test",
+  "types/empty_arguments",
+  "types/enum_as_input",
+  "types/enum_encoding",
+  "types/enum_inside_struct",
+  "types/evm_address",
+  "types/generics",
+  "types/identity",
+  "types/native_types",
+  "types/nested_structs",
+  "types/options",
+  "types/results",
+  "types/str_in_array",
+  "types/tuples",
+  "types/two_structs",
+  "types/type_inside_enum",
+  "types/vectors",
+
+  "storage/contract_storage_test",
+
+  "scripts/script_tuple_types",
+  "scripts/script_with_arguments",
+  "scripts/script_vectors",
+  "scripts/script_vector",
+  "scripts/script_require_from_contract",
+  "scripts/script_require",
+  "scripts/script_option_result_types",
+  "scripts/script_generic_types",
+  "scripts/basic_script",
+
+  "predicates/predicate_address",
+  "predicates/predicate_basic",
+  "predicates/predicate_vectors",
+  "predicates/predicate_vector",
+  "predicates/predicate_u64",
+  "predicates/predicate_tuple",
+  "predicates/predicate_structs",
+  "predicates/predicate_signatures",
+  "predicates/predicate_generics",
+  "predicates/predicate_enums",
+
+  "logs/contract_logs",
+  "logs/contract_logs_abi",
+  "logs/contract_with_contract_logs",
+  "logs/script_logs",
+  "logs/script_with_contract_logs",
+
+  "contracts/lib_contract_abi",
+  "contracts/contract_require_abi",
+  "contracts/lib_contract_caller",
+  "contracts/auth_testing_contract",
+  "contracts/auth_testing_abi",
+  "contracts/lib_contract",
+  "contracts/large_return_data",
+  "contracts/contract_test",
+  "contracts/library_test",
+  "contracts/liquidity_pool",
+  "contracts/msg_amount",
+  "contracts/multiple_read_calls",
+  "contracts/require",
+  "contracts/require_from_contract",
+  "contracts/revert_transaction_error",
+  "contracts/token_ops",
+  "contracts/transaction_block_height",
+
+]
