+++ scripts/test-projects/src/main.rs
@@ -2,15 +2,15 @@
 //!
 //! NOTE: This expects `forc`, `forc-fmt` and `cargo` to be available in `PATH`.
 
-use anyhow::Error;
 use test_projects::{
-    cli::parse_cli, display_failed, display_stats, run_command, types::ResultWriter,
+    cli::parse_cli, display_failed, display_info, display_stats, run_command, types::ResultWriter,
 };
 
 #[tokio::main]
-async fn main() -> Result<(), Error> {
+async fn main() {
     let mut result_writer = ResultWriter::new();
     let run_config = parse_cli();
+    display_info(&mut result_writer, &run_config);
 
     let (succeded, failed) = run_command(&mut result_writer, &run_config).await;
 
@@ -20,6 +20,4 @@ async fn main() -> Result<(), Error> {
     if !failed.is_empty() {
         std::process::exit(1);
     }
-
-    Ok(())
 }
